+++ frontend/src/components/common/HorizontalDivider.tsx
