+++ frontend/src/screens/AppLockScreen.tsx
@@ -1,7 +1,7 @@
 import React, {useState} from 'react';
 import {Dimensions, StyleSheet, Text, View} from 'react-native';
 
-import AppScreen from '../components/AppScreen';
+import AppScreen from '../components/layout/AppScreen';
 import {PASSWORD_LENGTH} from '../components/password/constants';
 import NumPad from '../components/password/NumPad';
 import PasswordInputArea from '../components/password/PasswordInputArea';
