+++ frontend/src/screens/auth/SignUpScreen.tsx
@@ -14,7 +14,7 @@ import {
 import * as yup from 'yup';
 
 import {signUp} from '../../api/authApi';
-import AppScreen from '../../components/AppScreen';
+import AppScreen from '../../components/layout/AppScreen';
 import type {AuthStackParamList} from '../../navigation/AuthStack';
 
 interface SignUpFormData {
