+++ frontend/src/screens/main/call-log/CallLogDetailScreen.tsx
@@ -4,8 +4,8 @@ import React from 'react';
 import {ScrollView, StyleSheet, Text, View} from 'react-native';
 
 import {RootStackParamList} from '../../../../App';
-import AppScreen from '../../../components/AppScreen';
-import Header from '../../../components/Header';
+import AppScreen from '../../../components/layout/AppScreen';
+import Header from '../../../components/layout/Header';
 import {formatKoreanDate, formatTime} from '../../../utils/date';
 
 type speakerType = 'ai' | 'user';
