+++ frontend/src/screens/main/settings/SetAppLockPasswordScreen.tsx
@@ -4,8 +4,8 @@ import React, {useState} from 'react';
 import {Alert, Dimensions, StyleSheet, Text, View} from 'react-native';
 
 import {RootStackParamList} from '../../../../App';
-import AppScreen from '../../../components/AppScreen';
-import Header, {HEADER_HEIGHT} from '../../../components/Header';
+import AppScreen from '../../../components/layout/AppScreen';
+import Header, {HEADER_HEIGHT} from '../../../components/layout/Header';
 import {PASSWORD_LENGTH} from '../../../components/password/constants';
 import NumPad from '../../../components/password/NumPad';
 import PasswordInputArea from '../../../components/password/PasswordInputArea';
